changed some typos, made the code even more complex, changed the input order, added some more input data

+++ main.py
@@ -16,6 +16,7 @@ def exit(reason):
 
 print("Mäuserechner\nBerechne, wie viele Mäuse nach einer bestimmten Anzahl an Zyklen leben würden.")
 
+
 #setup your own mice population
 try:
   youth = int(input("\nAnzahl der jungen Tiere: \n"))
@@ -23,12 +24,8 @@ try:
   old = int(input("\nAnzahl der alten Tiere: \n"))
   cycles = int(input("\nAnzahl der Zyklen: \n"))
 except:
-  exit("\nNope.")
+  exit("\n\nNope.")
 
-print("\nSoll die Anzahl der Tiere nach jedem Zyklus angezeigt werden?")
-showAnimalsAfterEveryCycle = input("y/n\n")
-if showAnimalsAfterEveryCycle != ("y" or "n"):
-  exit("WHY? just WHY?")
 
 print("\nSollen die einzelnen Werte verändert werden?")
 alterEverything = input("y/n\n")
@@ -39,7 +36,7 @@ if alterEverything == "y":
   boomersToOld = int(input("erwachsene Tiere zu alten Tieren: \n"))
   decimalPlaces = int(input("Nachkommastellen: \n"))
 elif alterEverything == "n":
-  print("\nNormaleinstellungen werden übernommen.")
+  print("Normaleinstellungen werden übernommen.")
   boomersToYouth = 4
   oldToYouth = 2
   youthToBoomers = 2
@@ -48,34 +45,51 @@ elif alterEverything == "n":
 else:
   exit("\n\nStop it. Get some help.")
 
-#open or create the csv file and run the mice simulation
-with open("table.csv", "w") as csvfile:
-  writer = csv.writer(csvfile, delimiter=";", quotechar='"', quoting=csv.QUOTE_MINIMAL)
+
+print("\nSoll die Anzahl der Tiere nach jedem Zyklus angezeigt werden?")
+showAnimalsAfterEveryCycle = input("y/n\n")
+try:
   
-  print("\n\nSimulation wird gestartet.")
+  if showAnimalsAfterEveryCycle != ("y" or "n"):
   
-  for i in range(cycles):
-    youthTemporarily = youth
-    boomersTemporarily = boomers
-    youth = round(((boomers * boomersToYouth) + (old * oldToYouth)),decimalPlaces)
-    boomers = round((youthTemporarily / youthToBoomers), decimalPlaces)
-    old = round((boomersTemporarily / boomersToOld), decimalPlaces)
     try:
-    writer.writerow([i + 1, youth, boomers, old])
+      
+      startSimulation = input("\nMit Enter die Simulation starten.")
+      if startSimulation == "":
+        #open or create the csv file and run the mice simulation
+        with open("table.csv", "w") as csvfile:
+          writer = csv.writer(csvfile, delimiter=";", quotechar='"', quoting=csv.QUOTE_MINIMAL)
+          
+          print("\nSimulation wird gestartet.")
+          
+          for i in range(cycles):
+            youthTemporarily = youth
+            boomersTemporarily = boomers
+            youth = round(((boomers * boomersToYouth) + (old * oldToYouth)),decimalPlaces)
+            boomers = round((youthTemporarily / youthToBoomers), decimalPlaces)
+            old = round((boomersTemporarily / boomersToOld), decimalPlaces)
+            try:
+              writer.writerow([i + 1, youth, boomers, old])
+            except:
+              exit("Zyklus " + str(i+1) + " fehlgeschlagen.")
+            else:
+              if (i+1) % 10 == 0:
+                print("Zyklus " + str(i+1) + " erfolgreich")
+          
+        
+            #if the user wants to, it is possible to see the current status after every successful cycle
+            if showAnimalsAfterEveryCycle == "y":
+              print("\nNach " + str(i + 1) + " Zyklen:")
+              print("Jungen Tiere: " + str(youth))
+              print("Erwachsene Tiere: " + str(boomers))
+              print("Alte Tiere: " + str(old))
+              print("Insgesamt " + str(youth + boomers + old) + " Tiere.")
     except:
-    exit("Zyklus " + str(i+1) + " fehlgeschlagen.")
-    else:
-    print("Zyklus " + str(i+1) + " erfolgreich")
+      exit("\n\nPRESS ENTER I SAID!")
+
+except:
+  exit("WHY? just WHY?")
   
-
-    #if the user wants to, it is possible to see the current status after every seccessful cycle
-    if showAnimalsAfterEveryCycle == "y":
-      print("\nNach " + str(i + 1) + " Zyklen:")
-      print("Jungen Tiere: " + str(youth))
-      print("Erwachsene Tiere: " + str(boomers))
-      print("Alte Tiere: " + str(old))
-      print("Insgesamt " + str(youth + boomers + old) + " Tiere.")
-
 #a review
 if showAnimalsAfterEveryCycle == "n":
   print("\nNach " + str(cycles) + " Zyklen:")
